Player Run animations hooked up, Attack and hit have bugs
Attack and hit are set to bool triggers which is likely causing the error as they return false the next frame. Using a timed Enumerator that is time synced to the animation speed will be the next approach to get the attack and hit animations to work.

Left Shoulder armour was parented to new player model. Will need artist to arrange armour on animated model.