Corrected scene name reference from "Level1" to "Level_1"
Several methods were not behaving as intended as they were looking for a scene that was not named with an underscore. All references to scene "Level_1" have been corrected.
Camera Depth has been assigned setting the gameplay camera with the highest priority and the temp camera with the lowest priority