Camera code clean up + bridge creation
The code for the camera was overly complex when the same effect could be achieved with a much more simple design.
The challenge of creating a semi endless bridge to run on and later generate waves of enemies has begun. There is a button that instantiates a large section of bridge to test performance impact of creating large sections during runtime. The current plan is to create a few sections during the introduction screen to be long enough that it consists of a one or more waves. During the results screen the player position will be sneakily reset. When the player exists the shop after purchasing upgrades they will "continue" their run with the wave count still run up but their progress on the bridge reset. When the last batch of waves is concluded there will be an event flag to whatever is resetting the player to disable and allow the player to reach the end.