Improved EnemySpawn, simplified EnemyFire, removed listener from Level_Manager
Whether or not the enemies can fire has been simplified in update while maintaining the toggle option.
Removed the need for a WaveSystem reference in Results Keeper.
Removed the EndWaveTrigger from Level_Manager.
ResetPlayerHealth heals the player before resetting them to avoid the death check repeating.
EnemySpawn's reference to MissZone is found in script to avoid the dependance on assignment in the inspector.