Weapon, Wave Logic, VFX bug fix
Tried to create a new weapon animation sequence where the hammer simply rotates in space and lerps between the player's hand and a target distance. This did not work out well, further work is needed to improve Hammer.
Wave system now correctly counts the waves however triggering the count to go up is complicated in its current state.
There was a bug with options that the Meteor VFX would play when the options menu was open but not during other gameplay-menus. I believe I have resolved this bug.