Enemy Wave Spawn bug testing
The array of locations that enemies can spawn within loses its reference during runtime. An overhaul of how the spawn locations are provided should solve this problem. Rather than grabbing references from the FirstWaveTrigger it is a better idea to tag all of the spawn positions with a tag of EnemySpawn and have each wave find all active objects with that tag to assign to the transform array of enemyspawn positions.