Restructured for more of a base defender style
The wave systems have been disabled in favor of a more simple approach.
EnemyProjectileManager communicates with new WaveUI to determine when the next wave is so that the enemies resume shooting.
BaseProjectile value from 8 to 4.
Removed attackCooldown in PlayerAttackHitBox as the second yield wait for new was creating bugs where the hammer would only destroy the projectiles some of the time