Weapon Attack Major Improvement
The default hammer attack animation now only travels as far as 20 ruler units.
The animation itself occurs in half a second and feels very smooth on the launch and the return.

The release trigger has been given a visual indicator of a magical circle that duplicates the item. This will become a scriptable object that can be upgraded.