Preview Buttons replaced
The preview buttons on the shop items that move that item info to the large display have been replaced by transparent images that cover nearly the entire item.
This allows anywhere on the item to be clicked on for the button to function. This will lead to a greatly improved player experince