Inverted Fast Zone
Fast Zone scales the speed of the projectiles in the first half of the map to accelerate them faster towards the player. This serves to build more of a frantic state as they appear to zoom at the player, the slow zone on the players side assures that the can still be reacted to fairly.